refactor: move templating logic to module
